requirements-pythonanywhere.txt
# Django Framework (PythonAnywhere Compatible)
Django>=5.2,<6.0

# Database adapter for PostgreSQL (production)
psycopg2-binary>=2.9.9

# Python environment variables (for production secrets)
python-decouple>=3.8

# Image processing (for profile photos and certificates)
Pillow>=10.0.0

# Development tools
django-debug-toolbar>=4.2.0

# Security
django-cors-headers>=4.3.0
# python-magic>=0.4.27  # Commented out - requires libmagic system library

# Production Web Server
gunicorn>=21.2.0

# Static Files Management (Production)
whitenoise>=6.6.0

# Database URL Parser (Production)
dj-database-url>=2.1.0

# Environment Variable Management
python-dotenv>=1.0.0
